Add files via upload
This Project showcases the contributions of women in STEM fields, with a special focus on African and Nigerian innovators. The app features:

Interactive Data Table: Browse and filter a dataset of trailblazing women in STEM.
Visual Insights: Charts highlighting contributions by country, field, and year.
Search & Filters: Find pioneers based on name, discipline, or achievements.
Built with Streamlit, Pandas, and Altair, this app celebrates the impact of women in STEM and their groundbreaking innovations.